requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.7
requests==2.19.0
urllib3==1.23
mysql-connector-python==8.0.11   
mysqlclient==1.3.12 
